Fix rngd bug that randomly keeps all CPUs busy after startup
RZ MPUs do not have a hardware random number generation module (/dev/hwrng), but that's the default setting for the systemd rngd service.
With that, rngd occupies the CPU resources to build up entropies during start-up.
Randomly, this process does not end properly and slows down the entire system significantly.
But if an available pseudorandom number generator is selected, rngd finishes properly always.

Signed-off-by: Nick <nick.cheng@avnet.com>